fix(review): compound bee key, sizes derivation comment, CLI flag guards, int annotation
Addresses code-reviewer P2 (React key uniqueness independent of sprite reuse)
and P3 notes from the feat/journal-hero review.

+++ apps/web/scripts/normalize_paper.py
@@ -19,7 +19,7 @@ from pathlib import Path
 from PIL import Image
 
 
-def paper_white(im: Image.Image, band: int = 48) -> tuple[float, float, float]:
+def paper_white(im: Image.Image, band: int = 48) -> tuple[int, int, int]:
     rgb = im.convert("RGB")
     w, h = rgb.size
     px = rgb.load()
